move tpi plugin to tpi package(test program interface)
